fix: 가게 관리 지도, history 지도 마커 수정

+++ app/src/main/java/com/kinopio/eatgo/domain/store/StoreHistory.kt
@@ -5,6 +5,6 @@ data class StoreHistory(
     val storeId: Integer,
     val openDate: String,
     val address: String,
-    val positionX: Integer,
-    val positionY: Integer
+    val positionX: Double,
+    val positionY: Double
 )
